Move the test code of sequencecheck to seqcheck_test.c. No functional change.
